nettoyage bin

+++ MPSoC/mutekh/tools/crossgen.mk
@@ -25,10 +25,11 @@
 #### LINE 25 IS HERE ####
 
 # Target architecture
-TARGET=mipsel
+#TARGET=mipsel
+TARGET=powerpc
 
 # Install PATH
-PREFIX=/opt/mutekh
+PREFIX=/users/enseig/genius/TTool/MPSoC/mutekh
 
 # Temp directory
 WORKDIR=/tmp/crossgen
@@ -39,7 +40,7 @@ BLDMAKE_OPTS= -j8
 # GNU Binutils
 binutils_VER_mipsel  = 2.20.1
 binutils_VER_powerpc = 2.20.1
-binutils_VER_arm     = 2.20.1
+binutils_VER_arm     = 2.23.2
 binutils_VER_i686    = 2.20.1
 binutils_VER_x86_64  = 2.20.1
 binutils_VER_nios2   = 2.20.1
@@ -47,21 +48,45 @@ binutils_VER_sparc   = 2.20.1
 binutils_VER_avr     = 2.20.1
 binutils_VER_lm32    = 2.20.1
 binutils_VER_microblaze = 2.20.1
+binutils_VER_avr32   = 2.22
 
 binutils_VER=$(binutils_VER_$(TARGET))
 binutils_CONF=
 
 # GNU Compiler
 gcc_VER_mipsel  = 4.5.2
+SUFFIX_mipsel   = unknown-elf
+
 gcc_VER_powerpc = 4.5.2
-gcc_VER_arm     = 4.5.2
+SUFFIX_powerpc  = unknown-elf
+
+gcc_VER_arm     = 4.8.2
+gcc_CONF_arm    = --with-arch=armv4t --with-fpu=vfp --with-float=soft
+SUFFIX_arm      = mutekh-eabi
+
 gcc_VER_i686    = 4.5.2
+SUFFIX_i868     = unknown-elf
+
 gcc_VER_x86_64  = 4.5.2
+SUFFIX_x86_64   = unknown-elf
+
 gcc_VER_nios2   = 4.4.4
+SUFFIX_nios2    = unknown-elf
+
 gcc_VER_sparc   = 4.5.2
+SUFFIX_nios2    = unknown-elf
+
 gcc_VER_avr     = 4.5.2
+SUFFIX_avr      = unknown-elf
+
 gcc_VER_lm32    = 4.5.2
+SUFFIX_lm32     = unknown-elf
+
 gcc_VER_microblaze = 4.5.2
+SUFFIX_microblaze  = unknown-elf
+
+gcc_VER_avr32   = 4.4.3
+SUFFIX_avr32    = unknown-elf
 
 gcc_VER=$(gcc_VER_$(TARGET))
 gcc_CONF=--enable-languages=c --disable-libssp --enable-multilib
@@ -72,19 +97,20 @@ gmp_VER=4.3.2
 mpc_VER=0.9
 
 # GNU Debugger
-gdb_VER_mipsel  = 7.2
-gdb_VER_powerpc = 7.2
-gdb_VER_arm     = 7.2
-gdb_VER_i686    = 7.2
-gdb_VER_x86_64  = 7.2
+gdb_VER_mipsel  = 7.4
+gdb_VER_powerpc = 7.4
+gdb_VER_arm     = 7.6.1
+gdb_VER_i686    = 7.4
+gdb_VER_x86_64  = 7.4
 gdb_VER_nios2   = 7.0
-gdb_VER_sparc   = 7.2
-gdb_VER_avr     = 7.2
-gdb_VER_lm32    = 7.2
-gdb_VER_microblaze = 7.3
+gdb_VER_sparc   = 7.4
+gdb_VER_avr     = 7.4
+gdb_VER_lm32    = 7.4
+gdb_VER_microblaze = 7.4
+gdb_VER_avr32   = 6.7.1
 
 gdb_VER=$(gdb_VER_$(TARGET))
-gdb_CONF=--with-python=no --disable-sim
+gdb_CONF=--with-python=no
 
 # Device Tree Compiler
 dtc_VER=1.2.0
@@ -107,21 +133,23 @@ unexport MAKEFLAGS
 unexport MFLAGS
 unexport MAKELEVEL
 
+SUFFIX=$(SUFFIX_$(TARGET))
+
 # packages configurations
 
 binutils_ARCHIVE=binutils-$(binutils_VER).tar.bz2
 binutils_URL=ftp://ftp.gnu.org/gnu/binutils/$(binutils_ARCHIVE)
-binutils_TESTBIN=bin/$(TARGET)-unknown-elf-as
+binutils_TESTBIN=bin/$(TARGET)-$(SUFFIX)-as
 
 gcc_ARCHIVE=gcc-$(gcc_VER).tar.bz2
 gcc_URL=ftp://ftp.gnu.org/gnu/gcc/gcc-$(gcc_VER)/$(gcc_ARCHIVE)
-gcc_TESTBIN=bin/$(TARGET)-unknown-elf-gcc
+gcc_TESTBIN=bin/$(TARGET)-$(SUFFIX)-gcc
 gcc_DEPS=binutils mpfr gmp mpc
 gcc_CONF+=--with-mpfr=$(PREFIX) --with-gmp=$(PREFIX) --with-mpc=$(PREFIX)
 
 gdb_ARCHIVE=gdb-$(gdb_VER).tar.bz2
-gdb_URL=ftp://ftp.gnu.org/gnu/gdb/gdb-$(gdb_VER)a.tar.bz2
-gdb_TESTBIN=bin/$(TARGET)-unknown-elf-gdb
+gdb_URL=ftp://ftp.gnu.org/gnu/gdb/gdb-$(gdb_VER).tar.bz2
+gdb_TESTBIN=bin/$(TARGET)-$(SUFFIX)-gdb
 
 mpfr_ARCHIVE=mpfr-$(mpfr_VER).tar.bz2
 mpfr_URL=ftp://ftp.gnu.org/gnu/mpfr/$(mpfr_ARCHIVE)
@@ -134,7 +162,7 @@ gmp_URL=ftp://ftp.gnu.org/gnu/gmp/$(gmp_ARCHIVE)
 gmp_TESTBIN=lib/libgmp.a
 
 mpc_ARCHIVE=mpc-$(mpc_VER).tar.gz
-mpc_URL=http://www.multiprecision.org/mpc/download/$(mpc_ARCHIVE)
+mpc_URL=http://www.multiprecision.org/downloads/$(mpc_ARCHIVE)
 mpc_TESTBIN=lib/libmpc.a
 mpc_DEPS=mpfr gmp
 mpc_CONF+=--with-mpfr=$(PREFIX) --with-gmp=$(PREFIX)
@@ -218,14 +246,14 @@ $$($(1)_TGZ): $$($(1)_STAMP)-wget
 
 $$($(1)_STAMP)-$$(TARGET)-patch: $$($(1)_DIR)
         # try to fetch a patch
-	wget $$(WGET_OPTS) $$(PATCH_URL)/$(1)-$$($(1)_VER)-$$(TARGET)-latest.diff.gz -O $$($(1)_PATCH).gz || rm -f $$($(1)_PATCH).gz
+	test -f $$($(1)_PATCH).gz || wget $$(WGET_OPTS) $$(PATCH_URL)/$(1)-$$($(1)_VER)-$$(TARGET)-latest.diff.gz -O $$($(1)_PATCH).gz || rm -f $$($(1)_PATCH).gz
         # test if a patch is available and apply
 	test ! -f $$($(1)_PATCH).gz || ( cd $$($(1)_DIR) ; cat $$($(1)_PATCH).gz | gunzip | patch -p 0 )
 	touch $$@
 
 $$($(1)_STAMP)-$$(TARGET)-conf: $$($(1)_DIR) $$($(1)_STAMP)-$$(TARGET)-patch $$($(1)_DEPS)
 	mkdir -p $$($(1)_BDIR)
-	( cd $$($(1)_BDIR) ; LD_LIBRARY_PATH=$$(PREFIX)/lib $$($(1)_DIR)/configure --disable-nls --prefix=$$(PREFIX) --target=$$(TARGET)-unknown-elf --disable-checking --disable-werror $$($(1)_CONF) ) && touch $$@
+	( cd $$($(1)_BDIR) ; LD_LIBRARY_PATH=$$(PREFIX)/lib $$($(1)_DIR)/configure MAKEINFO="makeinfo --version" --disable-nls --prefix=$$(PREFIX) --target=$$(TARGET)-$$(SUFFIX) --disable-checking --disable-werror $$($(1)_CONF) $$($(1)_CONF_$$(TARGET))  ) && touch $$@
 
 $$($(1)_STAMP)-$$(TARGET)-build: $$($(1)_STAMP)-$$(TARGET)-conf
 	LD_LIBRARY_PATH=$$(PREFIX)/lib make $$(BLDMAKE_OPTS) -C $$($(1)_BDIR) && touch $$@
@@ -258,7 +286,7 @@ $$($(1)_STAMP)-wget:
 
 $$($(1)_STAMP)-patch: $$($(1)_DIR)
         # try to fetch a patch
-	wget $$(WGET_OPTS) $$(PATCH_URL)/$(1)-$$($(1)_VER)-latest.diff.gz -O $$($(1)_PATCH).gz || rm -f $$($(1)_PATCH).gz
+	test -f $$($(1)_PATCH).gz || wget $$(WGET_OPTS) $$(PATCH_URL)/$(1)-$$($(1)_VER)-latest.diff.gz -O $$($(1)_PATCH).gz || rm -f $$($(1)_PATCH).gz
         # test is a patch is available and apply
 	test ! -f $$($(1)_PATCH).gz || ( cd $$($(1)_DIR) ; cat $$($(1)_PATCH).gz | gunzip | patch -p 0 )
 	touch $$@
